minor cleanups and bug fixes
  * fix bug in alm tasks state machine (cross enum)
  * fix bug in _Generic macro
  * rearrange some LINEAR11 code to avoid `powf` call
